moved singlecpu action header files from programs directory to actions directory
