+++ kernels/cpu/include/readdy/kernel/cpu/actions/CPUCompartments.h
@@ -30,7 +30,7 @@
 #ifndef READDY_CPUKERNEL_COMPARTMENTS_H
 #define READDY_CPUKERNEL_COMPARTMENTS_H
 
-#include <readdy/kernel/singlecpu/programs/SCPUCompartments.h>
+#include <readdy/kernel/singlecpu/actions/SCPUCompartments.h>
 #include <readdy/kernel/cpu/CPUKernel.h>
 
 namespace readdy {
