+++ kernels/singlecpu/src/SCPUActionFactory.cpp
@@ -25,12 +25,12 @@
 //
 
 #include <readdy/common/make_unique.h>
-#include <readdy/kernel/singlecpu/programs/SCPUActionFactory.h>
-#include <readdy/kernel/singlecpu/programs/SCPUEulerBDIntegrator.h>
-#include <readdy/kernel/singlecpu/programs/SCPUCalculateForces.h>
-#include <readdy/kernel/singlecpu/programs/SCPUReactionImpls.h>
-#include <readdy/kernel/singlecpu/programs/SCPUUpdateNeighborList.h>
-#include <readdy/kernel/singlecpu/programs/SCPUCompartments.h>
+#include <readdy/kernel/singlecpu/actions/SCPUActionFactory.h>
+#include <readdy/kernel/singlecpu/actions/SCPUEulerBDIntegrator.h>
+#include <readdy/kernel/singlecpu/actions/SCPUCalculateForces.h>
+#include <readdy/kernel/singlecpu/actions/SCPUReactionImpls.h>
+#include <readdy/kernel/singlecpu/actions/SCPUUpdateNeighborList.h>
+#include <readdy/kernel/singlecpu/actions/SCPUCompartments.h>
 
 namespace core_actions = readdy::model::actions;
